refactor: type 파일 위치 공통으로 변경
