+++ src/pages/feed-diaries/DiaryList/components/DiaryCardContent/index.tsx
@@ -1,4 +1,4 @@
-import type { Diary } from '../../type/diary';
+import type { Diary } from '@/shared/types/diary';
 import S from './style.module.css';
 
 interface Props {
