--- src/pages/feed-diaries/DiaryList/type/diary.ts
@@ -1,3 +1,0 @@
-import type { Database } from '@/shared/api/supabase/types';
-
-export type Diary = Database['public']['Tables']['diaries']['Row'];
