+++ src/pages/users/UserList/components/UserCard/index.tsx
@@ -2,7 +2,7 @@ import S from './style.module.css';
 import default_profile from '../../assets/default_profile.svg';
 import { GrFormNextLink } from 'react-icons/gr';
 import { useNavigate } from 'react-router-dom';
-import type { DbUser } from '@/pages/users/UserList/types/dbUser';
+import type { DbUser } from '@/shared/types/dbUser';
 
 const default_image = default_profile;
