+++ src/pages/users/UserList/components/UserList/index.tsx
@@ -1,4 +1,4 @@
-import type { DbUser } from '@/pages/users/UserList/types/dbUser';
+import type { DbUser } from '@/shared/types/dbUser';
 import UserCard from '../UserCard';
 import S from './style.module.css';
