+++ src/pages/users/UserList/hooks/useUserSearch.ts
@@ -1,4 +1,4 @@
-import type { DbUser } from '@/pages/users/UserList/types/dbUser';
+import type { DbUser } from '@/shared/types/dbUser';
 import { useMemo } from 'react';
 
 export const useUserSearch = (users: DbUser[], searchTerm: string) => {
